requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25

# Video Processing
yt-dlp==2024.1.26
moviepy==2.0.0.dev2

# Google APIs
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0

# Twilio (WhatsApp)
twilio==8.10.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
pillow==10.2.0
